\name{squared}
\alias{squared}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Square a Value}
\description{
Generic function to square a numeric value.}
\usage{
squared(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    a numeric value}
}
\value{
The squared value of x is computed.
}
\author{
Garret Binding, \email{binding@ipz.uzh.ch}
}
\examples{
squared(3)
}

